requirements.txt
streamlit == 1.36
streamlit-navigation-bar == 3.3.0
streamlit-pdf-viewer == 0.0.14
pandas == 2.2.2
numpy == 1.23.5
scikit-learn == 1.5.1
matplotlib == 3.9.1
xgboost == 2.1.0
pdfplumber == 0.11.2
transformers == 4.42.4
pypdf == 4.2.0
google-generativeai == 0.7.1
tensorflow == 2.12
openpyxl == 3.1.5

